-add academicWork and categoryBased
